refactor imgs into directory
