+++ tetrisgod/src/components/TodoPage.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import '../css/App.css';
-import logo from "../logo.svg";
+import logo from "../imgs/logo.svg";
 import Navbar from "./Navbar";
 import Footer from "./Footer";
